pyproject.toml
[project]
name = "pincer-x402-starter"
version = "0.1.2"
description = "Pincer x402 reference implementation with post-pay rebates"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    
    # x402 Protocol SDK (包含 Solana 支援)
    "x402>=0.3.0",
    
    # HTTP Clients
    "httpx>=0.28.0",
    "requests>=2.32.0",
    
    # Data Validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    
    # Configuration
    "python-dotenv>=1.0.0",
    
    # Blockchain - Solana (優先)
    "solana>=0.36.0",
    "solders>=0.21.0",
    
    # Blockchain - EVM (次要)
    "web3>=7.6.0",
    "eth-account>=0.13.0",
    
    # Database
    "aiosqlite>=0.20.0",
    
    # Utilities
    "python-dateutil>=2.9.0",
    
    # Internal Packages
    "pincer-sdk",
]

[project.optional-dependencies]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv.workspace]
members = ["packages/*"]

[tool.uv.sources]
pincer-sdk = { workspace = true }

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
]
